Move testing config functions into mtest package
