+++ metricbeat/module/traefik/mtest/testing.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package traefik
+package mtest
 
 import "os"
